refactoring of split and merge functions
- splitandmerge serves both single and full split&merge
- phonsplit and phonmerge modify an agent (population[[j]]) inplace
and return a logical indicating whether any modification (split or merge resp.) was made
- internally, both phonsplit and phonmerge call split_is_justified()

I had to use as.matrix() on agent$features for speed reasons, while I keep it as
data.table in order to modify it by ref (in perceive_token).

Performance of a full split and merge is almost identical with original and new code.